Fixes and code clean (formatting)
