+++ client/src/backgrounds/BioluminescenceBackground.tsx
@@ -24,11 +24,6 @@ interface Orb {
   b: number;
 }
 
-function hexToRgb(hex: string): [number, number, number] {
-  const c = hex.replace("#", "");
-  return [parseInt(c.slice(0, 2), 16), parseInt(c.slice(2, 4), 16), parseInt(c.slice(4, 6), 16)];
-}
-
 // Palette of deep-sea bioluminescent colors
 const PALETTE: [number, number, number][] = [
   [0, 220, 180], // teal
