+++ client/src/backgrounds/FishtankBackground.tsx
@@ -1,4 +1,5 @@
 import { memo, useEffect, useRef } from "react";
+import { hexToRgbTuple, lerp } from "@/lib/utils";
 
 interface Props {
   colorLeft: string;
@@ -11,26 +12,6 @@ interface Props {
   showSeaweed?: boolean;
   showCaustics?: boolean;
   showLightShafts?: boolean;
-}
-
-function hexToRgb(hex: string): [number, number, number] {
-  const c = hex.replace("#", "");
-  const full =
-    c.length === 3
-      ? c
-          .split("")
-          .map((x) => x + x)
-          .join("")
-      : c;
-  return [
-    parseInt(full.slice(0, 2), 16),
-    parseInt(full.slice(2, 4), 16),
-    parseInt(full.slice(4, 6), 16),
-  ];
-}
-
-function lerp(a: number, b: number, t: number) {
-  return a + (b - a) * t;
 }
 
 // ─── Fish ─────────────────────────────────────────────────────────────────────
@@ -293,9 +274,9 @@ export const FishtankBackground = memo(function FishtankBackground({
     resize();
     if (!preview) window.addEventListener("resize", resize);
 
-    const [lr, lg, lb] = hexToRgb(colorLeft);
-    const [mr, mg, mb] = hexToRgb(colorMiddle);
-    const [rr, rg, rb] = hexToRgb(colorRight);
+    const [lr, lg, lb] = hexToRgbTuple(colorLeft);
+    const [mr, mg, mb] = hexToRgbTuple(colorMiddle);
+    const [rr, rg, rb] = hexToRgbTuple(colorRight);
 
     // Palette: mix theme colors with vivid fish colors
     const palette: Array<[number, number, number]> = [
