+++ client/src/components/ErrorBoundary.tsx
@@ -40,7 +40,7 @@ export class ErrorBoundary extends React.Component<Props, State> {
         <div className="flex min-h-dvh flex-col items-center justify-center gap-4 px-6 text-center">
           <p className="text-lg font-medium text-white/80">Something went wrong on this page.</p>
           {this.state.error?.message && (
-            <p className="max-w-md font-mono text-sm break-words text-white/40">
+            <p className="max-w-md font-mono text-sm wrap-break-word text-white/40">
               {this.state.error.message}
             </p>
           )}
