+++ client/src/index.css
@@ -187,3 +187,190 @@
   border-bottom: max(env(safe-area-inset-bottom), 6px) solid rgba(245, 158, 11, 0.45);
   border-left: max(env(safe-area-inset-left), 6px) solid rgba(236, 72, 153, 0.45);
 }
+
+/* ── LavaLamp background keyframes ─────────────────────────────────────────── */
+@keyframes lava-blob-1 {
+  0% {
+    transform: translate(0%, 0%) scale(1);
+    border-radius: 62% 38% 56% 44% / 52% 62% 38% 48%;
+  }
+  20% {
+    transform: translate(7%, -22%) scale(1.14);
+    border-radius: 38% 62% 44% 56% / 62% 38% 56% 44%;
+  }
+  45% {
+    transform: translate(-9%, 15%) scale(0.88);
+    border-radius: 56% 44% 62% 38% / 44% 56% 48% 52%;
+  }
+  70% {
+    transform: translate(14%, -8%) scale(1.1);
+    border-radius: 44% 56% 38% 62% / 56% 44% 62% 38%;
+  }
+  100% {
+    transform: translate(0%, 0%) scale(1);
+    border-radius: 62% 38% 56% 44% / 52% 62% 38% 48%;
+  }
+}
+@keyframes lava-blob-2 {
+  0% {
+    transform: translate(0%, 0%) scale(1);
+    border-radius: 46% 54% 52% 48% / 62% 38% 52% 48%;
+  }
+  28% {
+    transform: translate(-12%, 18%) scale(1.18);
+    border-radius: 62% 38% 44% 56% / 38% 62% 56% 44%;
+  }
+  58% {
+    transform: translate(9%, -14%) scale(0.84);
+    border-radius: 38% 62% 56% 44% / 56% 44% 38% 62%;
+  }
+  100% {
+    transform: translate(0%, 0%) scale(1);
+    border-radius: 46% 54% 52% 48% / 62% 38% 52% 48%;
+  }
+}
+@keyframes lava-blob-3 {
+  0% {
+    transform: translate(0%, 0%) scale(1);
+    border-radius: 52% 48% 38% 62% / 46% 56% 62% 38%;
+  }
+  38% {
+    transform: translate(15%, -24%) scale(1.24);
+    border-radius: 38% 62% 62% 38% / 62% 38% 38% 62%;
+  }
+  68% {
+    transform: translate(-11%, 10%) scale(0.9);
+    border-radius: 62% 38% 38% 62% / 38% 62% 62% 38%;
+  }
+  100% {
+    transform: translate(0%, 0%) scale(1);
+    border-radius: 52% 48% 38% 62% / 46% 56% 62% 38%;
+  }
+}
+@keyframes lava-blob-4 {
+  0% {
+    transform: translate(0%, 0%) scale(1);
+    border-radius: 56% 44% 52% 48% / 52% 48% 44% 56%;
+  }
+  33% {
+    transform: translate(-16%, 12%) scale(1.13);
+    border-radius: 44% 56% 62% 38% / 62% 38% 56% 44%;
+  }
+  63% {
+    transform: translate(10%, -18%) scale(0.83);
+    border-radius: 62% 38% 44% 56% / 38% 62% 52% 48%;
+  }
+  100% {
+    transform: translate(0%, 0%) scale(1);
+    border-radius: 56% 44% 52% 48% / 52% 48% 44% 56%;
+  }
+}
+@keyframes lava-blob-5 {
+  0% {
+    transform: translate(0%, 0%) scale(1);
+    border-radius: 52% 48% 56% 44% / 56% 44% 52% 48%;
+  }
+  42% {
+    transform: translate(7%, 25%) scale(1.2);
+    border-radius: 44% 56% 46% 54% / 44% 56% 56% 44%;
+  }
+  100% {
+    transform: translate(0%, 0%) scale(1);
+    border-radius: 52% 48% 56% 44% / 56% 44% 52% 48%;
+  }
+}
+@keyframes lava-blob-6 {
+  0% {
+    transform: translate(0%, 0%) scale(1);
+    border-radius: 62% 38% 52% 48% / 52% 62% 44% 56%;
+  }
+  48% {
+    transform: translate(-12%, -22%) scale(1.26);
+    border-radius: 38% 62% 62% 38% / 62% 38% 56% 44%;
+  }
+  100% {
+    transform: translate(0%, 0%) scale(1);
+    border-radius: 62% 38% 52% 48% / 52% 62% 44% 56%;
+  }
+}
+@keyframes highlight-sweep {
+  0% {
+    opacity: 0.1;
+    transform: translateX(-60%) skewX(-18deg);
+  }
+  40% {
+    opacity: 0.25;
+    transform: translateX(0%) skewX(-18deg);
+  }
+  100% {
+    opacity: 0.1;
+    transform: translateX(60%) skewX(-18deg);
+  }
+}
+
+/* ── Sakura background keyframes ────────────────────────────────────────────── */
+@keyframes sakura-fall {
+  0% {
+    transform: translateY(-60px) translateX(0px) rotate(0deg) scale(1);
+    opacity: 0;
+  }
+  5% {
+    opacity: 1;
+  }
+  90% {
+    opacity: 0.85;
+  }
+  100% {
+    transform: translateY(110vh) translateX(var(--drift)) rotate(var(--rot-end)) scale(0.7);
+    opacity: 0;
+  }
+}
+@keyframes sakura-sway {
+  0% {
+    margin-left: 0px;
+  }
+  25% {
+    margin-left: var(--sway);
+  }
+  75% {
+    margin-left: calc(var(--sway) * -0.6);
+  }
+  100% {
+    margin-left: 0px;
+  }
+}
+
+/* ── Storm background keyframes ─────────────────────────────────────────────── */
+@keyframes cloud-drift-1 {
+  0% {
+    transform: translateX(-4%);
+  }
+  50% {
+    transform: translateX(4%);
+  }
+  100% {
+    transform: translateX(-4%);
+  }
+}
+@keyframes cloud-drift-2 {
+  0% {
+    transform: translateX(5%);
+  }
+  50% {
+    transform: translateX(-3%);
+  }
+  100% {
+    transform: translateX(5%);
+  }
+}
+@keyframes cloud-drift-3 {
+  0% {
+    transform: translateX(-2%);
+  }
+  50% {
+    transform: translateX(6%);
+  }
+  100% {
+    transform: translateX(-2%);
+  }
+}
